Update README.md - add removing instructions
Add a Paragraph to explain, how the patch could be removed, from installation.
I'm not sure, if this commands works for everyone. 
And maybe we should add an information that the patch did not do any sustainable changes, so it could be reverted and everything is like it was before?